# Reuse call stack when invoking filter